Move last bits out of observables file
